feat: hook single reponsability on add asset
