+++ frontend/pages/home/components/ICRC/asset/AddAssetManual.tsx
@@ -10,7 +10,7 @@ import { CustomButton } from "@components/button";
 import { useTranslation } from "react-i18next";
 import { IconTypeEnum } from "@/const";
 import { Asset } from "@redux/models/AccountModels";
-import { ChangeEvent, Dispatch, SetStateAction, useState } from "react";
+import { ChangeEvent, useEffect, useState } from "react";
 import { Principal } from "@dfinity/principal";
 import { LoadingLoader } from "@components/loader";
 import { AccountHook } from "@pages/hooks/accountHook";
@@ -18,37 +18,22 @@ import { getAssetDetails } from "@pages/home/helpers/icrc";
 import { db } from "@/database/db";
 import { Contact } from "@redux/models/ContactsModels";
 import { getAssetIcon } from "@/utils/icons";
-import { AssetMutationAction, setAssetMutation, setAssetMutationAction } from "@redux/assets/AssetReducer";
+import {
+  AssetMutationAction,
+  AssetMutationResult,
+  setAccordionAssetIdx,
+  setAssetMutation,
+  setAssetMutationAction,
+  setAssetMutationResult,
+  setSelectedAsset,
+} from "@redux/assets/AssetReducer";
 import { useAppDispatch, useAppSelector } from "@redux/Store";
-import { assetMutateInitialState } from "@pages/home/hooks/useAssetMutate";
+import useAssetMutate, { assetMutateInitialState } from "@pages/home/hooks/useAssetMutate";
 
-interface AddAssetManualProps {
-  errToken: string;
-  setErrToken: Dispatch<SetStateAction<string>>;
-  errIndex: string;
-  setErrIndex: Dispatch<SetStateAction<string>>;
-  validToken: boolean;
-  setValidToken: Dispatch<SetStateAction<boolean>>;
-  newAsset: Asset;
-  setNewAsset: Dispatch<SetStateAction<Asset>>;
-  asset: Asset | undefined;
-  addAssetToData(): void;
-}
-
-const AddAssetManual = (props: AddAssetManualProps) => {
-  const { assetAction } = useAppSelector((state) => state.asset.mutation);
-  const {
-    errToken,
-    setErrToken,
-    errIndex,
-    setErrIndex,
-    validToken,
-    setValidToken,
-    newAsset,
-    setNewAsset,
-    asset,
-    addAssetToData,
-  } = props;
+const AddAssetManual = () => {
+  const { assetAction, assetMutated } = useAppSelector((state) => state.asset.mutation);
+  const { assets } = useAppSelector((state) => state.asset);
+  const { newAsset, setNewAsset, setErrToken, errToken } = useAssetMutate();
 
   const { t } = useTranslation();
   const dispatch = useAppDispatch();
@@ -59,13 +44,45 @@ const AddAssetManual = (props: AddAssetManualProps) => {
   const [tested, setTested] = useState(false);
   const [errShortDec, serErrShortDec] = useState(false);
   const [validIndex, setValidIndex] = useState(false);
+  const [errIndex, setErrIndex] = useState("");
+  const [validToken, setValidToken] = useState(false);
+  const isUpdate = assetAction === AssetMutationAction.UPDATE;
+
+  useEffect(() => {
+    if (assetMutated) {
+      setNewAsset({
+        address: assetMutated.address,
+        symbol: assetMutated.symbol,
+        name: assetMutated.name,
+        tokenName: assetMutated.tokenName,
+        tokenSymbol: assetMutated.tokenSymbol,
+        decimal: assetMutated.decimal,
+        shortDecimal: assetMutated.shortDecimal,
+        subAccounts: assetMutated.subAccounts.map((ast) => {
+          return {
+            name: ast.name,
+            sub_account_id: ast.sub_account_id,
+            amount: ast.amount,
+            currency_amount: ast.currency_amount,
+            address: ast.address,
+            decimal: ast.decimal,
+            symbol: ast.symbol,
+            transaction_fee: ast.transaction_fee,
+          };
+        }),
+        index: assetMutated.index,
+        sortIndex: assetMutated.sortIndex,
+        supportedStandards: assetMutated.supportedStandards,
+      });
+    }
+  }, [assetMutated]);
 
   return (
     <div className="flex flex-col items-start justify-start w-full">
-      {assetAction === AssetMutationAction.UPDATE && asset ? (
+      {isUpdate ? (
         <div className="flex flex-col items-center justify-start w-full p-2">
-          {getAssetIcon(IconTypeEnum.Enum.ASSET, asset.tokenSymbol, asset.logo)}
-          <p className="mt-2 text-lg font-bold">{`${asset.name} - ${asset.symbol}`}</p>
+          {getAssetIcon(IconTypeEnum.Enum.ASSET, newAsset.tokenSymbol, newAsset.logo)}
+          <p className="mt-2 text-lg font-bold">{`${newAsset.name} - ${newAsset.symbol}`}</p>
         </div>
       ) : (
         <div className="flex flex-row items-start justify-start w-full p-2 border rounded-lg border-SelectRowColor bg-SelectRowColor/10">
@@ -82,8 +99,8 @@ const AddAssetManual = (props: AddAssetManualProps) => {
           sizeInput={"medium"}
           sufix={<CustomCopy size={"small"} copyText={newAsset.address} side="left" align="center" />}
           intent={"secondary"}
-          disabled={asset ? true : false}
-          inputClass={asset ? "opacity-40" : ""}
+          disabled={isUpdate ? true : false}
+          inputClass={isUpdate ? "opacity-40" : ""}
           placeholder="Ledger Principal"
           compOutClass=""
           value={newAsset.address}
@@ -110,7 +127,7 @@ const AddAssetManual = (props: AddAssetManualProps) => {
         {validIndex && <p className="text-sm text-left text-slate-color-info">{t("index.validation.msg")}</p>}
       </div>
 
-      {!asset && (
+      {!isUpdate && (
         <div className="flex justify-end w-full">
           <CustomButton
             intent={newAsset.address.length > 5 ? "success" : "deny"}
@@ -131,7 +148,7 @@ const AddAssetManual = (props: AddAssetManualProps) => {
           compOutClass=""
           value={newAsset.symbol}
           onChange={onChangeSymbol}
-          disabled={!asset && !tested}
+          disabled={!isUpdate && !tested}
         />
       </div>
 
@@ -144,7 +161,7 @@ const AddAssetManual = (props: AddAssetManualProps) => {
           compOutClass=""
           value={newAsset.name}
           onChange={onChangeName}
-          disabled={!asset && !tested}
+          disabled={!isUpdate && !tested}
         />
       </div>
 
@@ -160,12 +177,12 @@ const AddAssetManual = (props: AddAssetManualProps) => {
         />
       </div>
 
-      <div className={`flex flex-row justify-start items-center ${asset ? "w-[85%]" : "w-full"} gap-2`}>
+      <div className={`flex flex-row justify-start items-center ${isUpdate ? "w-[85%]" : "w-full"} gap-2`}>
         <div className="flex flex-col items-start w-full mb-3">
           <p className="opacity-60">{t("token.decimal")}</p>
           <CustomInput
             sizeInput={"medium"}
-            inputClass={asset ? "opacity-40" : ""}
+            inputClass={isUpdate ? "opacity-40" : ""}
             intent={"secondary"}
             placeholder=""
             disabled={true}
@@ -175,7 +192,7 @@ const AddAssetManual = (props: AddAssetManualProps) => {
             onChange={onChangeDecimal}
           />
         </div>
-        {asset && (
+        {isUpdate && (
           <div className="flex flex-col items-start w-full mb-3">
             <p className="opacity-60 text-md">{t("short.form.limit")}</p>
             <CustomInput
@@ -198,9 +215,9 @@ const AddAssetManual = (props: AddAssetManualProps) => {
           </CustomButton>
         )}
         <CustomButton
-          intent={(asset ? newAsset.address.length > 5 : tested) ? "accept" : "deny"}
+          intent={(isUpdate ? newAsset.address.length > 5 : tested) ? "accept" : "deny"}
           onClick={onSave}
-          disabled={asset ? newAsset.address.length <= 5 : !tested}
+          disabled={isUpdate ? newAsset.address.length <= 5 : !tested}
         >
           {t("save")}
         </CustomButton>
@@ -333,7 +350,8 @@ const AddAssetManual = (props: AddAssetManualProps) => {
   }
 
   async function onSave() {
-    if (asset) {
+    if (isUpdate) {
+      // INFO: saving an updated asset
       if (newAsset.shortDecimal === "") {
         serErrShortDec(true);
         return;
@@ -350,7 +368,7 @@ const AddAssetManual = (props: AddAssetManualProps) => {
           const newDoc = {
             ...contact,
             assets: contact.assets.map((currentAsset) => {
-              if (currentAsset.tokenSymbol === asset?.tokenSymbol) {
+              if (currentAsset.tokenSymbol === newAsset?.tokenSymbol) {
                 affected = true;
                 return { ...currentAsset, symbol: newAsset.symbol };
               } else return currentAsset;
@@ -367,9 +385,9 @@ const AddAssetManual = (props: AddAssetManualProps) => {
         );
       }, 0);
 
-      const asset = await db().getAsset(newAsset.address);
+      const assetDB = await db().getAsset(newAsset.address);
 
-      if (asset) {
+      if (assetDB) {
         // INFO: update an asset
         const updatedFull: Asset = {
           ...newAsset,
@@ -379,7 +397,7 @@ const AddAssetManual = (props: AddAssetManualProps) => {
               ? Number(newAsset.decimal).toFixed(0)
               : Number(newAsset.shortDecimal).toFixed(0),
         };
-        await db().updateAsset(asset.address, updatedFull, { sync: true });
+        await db().updateAsset(assetDB.address, updatedFull, { sync: true });
       }
 
       setNewAsset(assetMutateInitialState);
@@ -387,6 +405,50 @@ const AddAssetManual = (props: AddAssetManualProps) => {
       dispatch(setAssetMutation(undefined));
       dispatch(setAssetMutationAction(AssetMutationAction.NONE));
     } else if (await onTest(false)) addAssetToData();
+  }
+
+  async function addAssetToData() {
+    if (isAssetAdded(newAsset.address)) {
+      setErrToken(t("adding.asset.already.imported"));
+      setValidToken(false);
+      return;
+    }
+
+    try {
+      dispatch(setAssetMutationResult(AssetMutationResult.ADDING));
+
+      const idxSorting = assets.length > 0 ? [...assets].sort((a, b) => b.sortIndex - a.sortIndex) : [];
+      const sortIndex = (idxSorting.length > 0 ? idxSorting[0]?.sortIndex : 0) + 1;
+
+      const updatedAsset = await getAssetDetails({
+        canisterId: newAsset.address,
+        includeDefault: true,
+        customName: newAsset.name,
+        customSymbol: newAsset.symbol,
+        supportedStandard: newAsset.supportedStandards,
+        sortIndex,
+      });
+
+      const assetToSave: Asset = { ...newAsset, ...updatedAsset, sortIndex };
+      await db().addAsset(assetToSave, { sync: true });
+
+      dispatch(setAssetMutationResult(AssetMutationResult.ADDED));
+      dispatch(setSelectedAsset(assetToSave));
+      dispatch(setAccordionAssetIdx([assetToSave.symbol]));
+    } catch (error) {
+      console.error("Error adding asset", error);
+      dispatch(setAssetMutationResult(AssetMutationResult.FAILED));
+    } finally {
+      setTimeout(() => {
+        dispatch(setAssetMutationResult(AssetMutationResult.NONE));
+        dispatch(setAssetMutationAction(AssetMutationAction.NONE));
+        dispatch(setAssetMutation(undefined));
+      }, 3000);
+    }
+  }
+
+  function isAssetAdded(address: string) {
+    return assets.find((asst: Asset) => asst.address === address) ? true : false;
   }
 };
